Add console logging and enhance view calls control
Added another button to view calls for only the current user or all users. This button in combination with the "View calls from today" button should allow for better visibility of calls in the database.

If the "Only your calls" button is not checked, all calls are shown and your calls are highlighted in green.